etl/requirements.txt
requests
pydantic
python-graphql-client==0.3.0
python-gitlab
rdflib>=6.1.1
SPARQLWrapper>=1.8.5
PyYAML
# prance[osv,icu,cli]
prance
obonet
PyShExC
# d2s-cli@git+https://github.com/MaastrichtU-IDS/d2s-cli.git@master
# git+https://github.com/MaastrichtU-IDS/d2s-cli.git@develop
# d2s